получает комментарии с сервера
