+++ src/components/popup/components/comments/comment-markup.js
@@ -1,3 +1,5 @@
+import {getCommentDate} from "../../../../tools/utils/utils";
+
 const createCommentTemplate = ({id, text, name, date, emoji}) => {
   return (
     `<li data-id="${id}" class="film-details__comment">
@@ -8,7 +10,7 @@ const createCommentTemplate = ({id, text, name, date, emoji}) => {
         <p class="film-details__comment-text">${text}</p>
         <p class="film-details__comment-info">
           <span class="film-details__comment-author">${name}</span>
-          <span class="film-details__comment-day">${date}</span>
+          <span class="film-details__comment-day">${getCommentDate(date)}</span>
           <button class="film-details__comment-delete">Delete</button>
         </p>
       </div>
