+++ src/components/popup/components/comments/create-comment-list.js
@@ -1,12 +1,12 @@
 import {createCommentsMurkup} from "./comments-list-markup";
 import Comment from "../../comments";
 
-const createCommentsList = (comments) => {
-  const listMarkup = comments.map((comment) => new Comment(comment).getTemplate()).join(`\n`);
+// const createCommentsList = (comments) => {
+//   const listMarkup = comments.map((comment) => new Comment(comment).getTemplate()).join(`\n`);
 
-  return (
-    createCommentsMurkup(listMarkup, comments)
-  );
-};
+//   return (
+//     createCommentsMurkup(listMarkup, comments)
+//   );
+// };
 
-export {createCommentsList};
+// export {createCommentsList};
