+++ src/tools/utils/utils.js
@@ -38,12 +38,8 @@ export const getBoolean = () => Math.random() > 0.5;
 export const getCommentDate = (commentDate) => {
   const today = new Date();
 
-  if (commentDate.getDate() === today.getDate()) {
+  if (commentDate === today) {
     return moment(commentDate).fromNow();
-  } else if ((today.getFullYear() === commentDate.getFullYear())
-    && (today.getMonth() - commentDate.getMonth())
-    && ((today.getDate() - commentDate.getDate()) <= WEEK)) {
-    return `${(today.getDate() - commentDate.getDate())} days ago`;
   } else {
     return formatDate(commentDate, FormatDate.COMMENT_DATE);
   }
